Renamed and moved utilities file
