+++ django_project/ford3/views/qualification_wizard.py
@@ -5,7 +5,7 @@ from django.urls import reverse
 from django.forms.models import model_to_dict
 from django.template.defaulttags import register
 from formtools.wizard.views import CookieWizardView
-from ford3.forms.form_utlities import get_form_identifier_list_from_keys
+from ford3.views.wizard_utilities import get_form_identifier_list_from_keys
 from ford3.models.qualification import Qualification
 from ford3.models.requirement import Requirement
 from ford3.models.subject import Subject
